cluster: move TryIntoTimelyConfig trait into mz_cluster_client
Removes deps:
- mz-cluster -> mz-compute-client
- mz-cluster -> mz-storage-client
