+++ src/cluster/Cargo.toml
@@ -17,7 +17,6 @@ differential-dataflow = "0.12.0"
 futures = "0.3.25"
 mz-build-info = { path = "../build-info" }
 mz-cluster-client = { path = "../cluster-client" }
-mz-compute-client = { path = "../compute-client" }
 mz-expr = { path = "../expr" }
 mz-ore = { path = "../ore", features = ["async", "tracing_"] }
 mz-persist-client = { path = "../persist-client" }
@@ -25,7 +24,6 @@ mz-persist-types = { path = "../persist-types" }
 mz-pid-file = { path = "../pid-file" }
 mz-repr = { path = "../repr" }
 mz-service = { path = "../service" }
-mz-storage-client = { path = "../storage-client" }
 mz-timely-util = { path = "../timely-util" }
 once_cell = "1.16.0"
 prometheus = { version = "0.13.3", default-features = false }
